refactor(ark-cli): reorganize the code into folders
