+++ ark-cli/Cargo.toml
@@ -9,7 +9,7 @@ bench = false
 
 [dependencies]
 tokio = { version = "1.35.1", features = ["full"] }
-clap = { version = "3.0.10", features = ["derive"] }
+clap = { version = "4.5", features = ["derive"] }
 env_logger = "0.9.0"
 fs_extra = "1.2.0"
 home = "0.5.3"
