+++ ark-cli/src/commands/link/mod.rs
@@ -1,0 +1,12 @@
+use clap::Subcommand;
+
+pub mod create;
+mod load;
+mod utils;
+
+/// Available commands for the `link` subcommand
+#[derive(Subcommand, Debug)]
+pub enum Link {
+    Create(create::Create),
+    Load(load::Load),
+}
