+++ ark-cli/src/commands/storage/mod.rs
@@ -1,0 +1,9 @@
+use clap::Subcommand;
+
+mod list;
+
+/// Available commands for the `storage` subcommand
+#[derive(Subcommand, Debug)]
+pub enum Storage {
+    List(list::List),
+}
