--- ark-cli/src/models/entry.rs
@@ -1,9 +1,0 @@
-use clap::Parser;
-
-#[derive(Parser, Debug, Clone, Copy, PartialEq, Eq)]
-pub enum EntryOutput {
-    Link,
-    Id,
-    Path,
-    Both,
-}
