--- ark-cli/src/models/sort.rs
@@ -1,7 +1,0 @@
-use clap::Parser;
-
-#[derive(Parser, Debug, clap::ValueEnum, Clone)]
-pub enum Sort {
-    Asc,
-    Desc,
-}
